Rename StockLineage to ToolLineage and merge stock.py contents into interface.py
